VIC-12248 SDK reserve control (#147)
These changes allow the SDK user to keep Vector still between scripts, or in a single script between instances of behavior control.

The primary way to use this is as a module script, either from the command-line or using Mac & Windows scripts in the sdk/examples/scripts folder that can be double-clicked to reserve control from the Finder or Windows Explorer.  There is also sample code in the documentation for how to use a ReserveBehaviorControl object in a script.